tests/wpt/css-tests/css-writing-modes-3_dev/xhtml1print/unicode-bidi-parsing-002.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Writing Modes: parsing unicode-bidi: isolate, isolate-override, plaintext</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii" />
<link href="https://drafts.csswg.org/css-writing-modes-3/#unicode-bidi" rel="help" />
<meta content="This test asserts the parser and getComputedStyle works correctly for the unicode-bidi: isolate, isolate-override, plaintext." name="assert" />
<meta content="dom" name="flags" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head><body><div data-expected="isolate" style="unicode-bidi: isolate"></div>
<div data-expected="isolate-override" style="unicode-bidi: isolate-override"></div>
<div data-expected="plaintext" style="unicode-bidi: plaintext"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).unicodeBidi;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
</body></html>
